==> 4_1.py <==
# -*- coding: utf-8 -*-
"""
Created on Tue Oct 30 10:28:39 2018


"""
#数据规范化
import pandas as pd


datafile = 'F:\spyder\datamining\chapter4\demo\data\discretization_data.xls' #参数初始化
data = pd.read_excel(datafile) #读取数据
data = data[u'肝气郁结证型系数'].copy()
k = 4

d1 = pd.cut(data, k, labels = range(k)) #等宽离散化，各个类比依次命名为0,1,2,3

#等频率离散化
w = [1.0*i/k for i in range(k+1)]
w = data.describe(percentiles = w)[4:4+k+1] #使用describe函数自动计算分位数
w[0] = w[0]*(1-1e-10)#设置保留多少位小数，原始的w[0]有3为小数，该语句后该w[0]有13位小数
d2 = pd.cut(data, w, labels = range(k))

#根据聚类结果进行离散化
from sklearn.cluster import KMeans #引入KMeans
kmodel = KMeans(n_clusters = 4) #建立模型
kmodel.fit(data.values.reshape((len(data), 1))) #训练模型
c = pd.DataFrame(kmodel.cluster_centers_).sort_values(0) #输出聚类中心，并且排序（默认是随机序的）
w = c.rolling(2).mean().iloc[1:]#从第一行开始算起，因为第0行的值为NAN（相邻两项求中点，作为边界点）
w = [0] + list(w[0]) + [data.max()]
d3 = pd.cut(data, w, labels = range(k))

def cluster_plot(d, k): #自定义作图函数来显示聚类结果
  import matplotlib.pyplot as plt
  plt.rcParams['font.sans-serif'] = ['SimHei'] #用来正常显示中文标签
  plt.rcParams['axes.unicode_minus'] = False #用来正常显示负号
  
  plt.figure(figsize = (8, 3))
  for j in range(0, k):
    plt.plot(data[d==j], [j for i in d[d==j]], 'o')
  
  plt.ylim(-0.5, k-0.5)
  return plt

cluster_plot(d1, k).show()

cluster_plot(d2, k).show()
cluster_plot(d3, k).show()
