==> 5_0.py <==
# -*- coding: utf-8 -*-
"""
Created on Thu Nov  1 15:13:14 2018


"""

import pandas as pd

#参数初始化
inputfile = 'F:\\spyder\\datamining\\chapter5\\demo\\data\\consumption_data.xls'
outputfile = 'F:\\spyder\\datamining\\chapter5\\data_type.xls'#保存结果的文件名

k = 3 #聚类的类别
iteration = 500 #聚类最大的循环次数
data = pd.read_excel(inputfile, index_col='Id')
data_zs = 1.0*(data - data.mean())/data.std()#数据标准化(零-均值规范化)

from sklearn.cluster import KMeans
model = KMeans(n_clusters=k, max_iter=iteration)
model.fit(data_zs)  

#简单打印结果
r1 = pd.Series(model.labels_).value_counts() 
r2 = pd.DataFrame(model.cluster_centers_)#聚类中心
r = pd.concat([r2, r1], axis=1)#横向连接（0代表纵向连接），得到聚类中心对应类别下的数目
r.columns = list(data.columns) + [u'类别数目']#重命名表头
print(r)        
        
#详细给出原始数据及其类别
r = pd.concat([data, pd.Series(model.labels_, index=data.index)], axis=1) #详细输出每个样本对应的类别
r.columns = list(data.columns) + [u'聚类类别'] #重命名表头
r.to_excel(outputfile) 

'''def density_plot(data): #自定义作图函数
  import matplotlib.pyplot as plt
  plt.rcParams['font.sans-serif'] = ['SimHei'] #用来正常显示中文标签
  plt.rcParams['axes.unicode_minus'] = False #用来正常显示负号
  p = data.plot(kind='kde', linewidth = 2, subplots = True, sharex = False)
  [p[i].set_ylabel(u'密度') for i in range(k)]
  plt.legend()
  return plt

pic_output = 'F:\\spyder\\datamining\\chapter5\\pd_' #概率密度图文件名前缀
for i in range(k):
  density_plot(data[r[u'聚类类别']==i]).savefig(u'%s%s.png' %(pic_output, i))'''
  
              
from sklearn.manifold import TSNE

tsne = TSNE()
tsne.fit_transform(data_zs) #进行数据将维
tsne = pd.DataFrame(tsne.embedding_, index=data_zs.index) #转换数据格式

import matplotlib.pyplot as plt
plt.rcParams['font.sans-serif'] = ['SimHei'] #用来正常显示中文标签
plt.rcParams['axes.unicode_minus'] = False #用来正常显示负号            
              
#不同类别使用不同颜色和样式绘图
d = tsne[r[u'聚类类别'] == 0]
plt.plot(d[0], d[1], 'r.')        
d = tsne[r[u'聚类类别'] == 1]
plt.plot(d[0], d[1], 'go')               
d = tsne[r[u'聚类类别'] == 2]
plt.plot(d[0], d[1], 'b*')   
plt.show()
